requirements.txt
streamlit
pandas
openpyxl
psycopg2-binary>=2.2.0
openpyxl>=3.1.5
